Development/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Presidio
presidio-analyzer==2.2.33
presidio-anonymizer==2.2.33

# Encryption
pycryptodome==3.19.0

# Document processing
PyPDF2==3.0.1
reportlab==4.0.7
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.3

# Web scraping and search
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0

# Image OCR
pytesseract==0.3.10
Pillow==10.1.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
